Add fallback for geocoding service
The first thing to know is that the geocoder for this project is not
meant to be exact. It is supposed to visualize clusters of members in a
particular region of about 25 miles. With that in mind, this PR attempts
to fill in some of the gaps where the census data is incomplete for a
particular address.

It retries with:

- 100 Main St with no zip code
- 100 2nd Ave with no zip code

If neither of those return a value, the coordinates revert to `0,0`