Add Directory Collections
This is the first step in making Member Directory more versatile outside of the college undergraduate/alumnus type of membership organization. It does this by making the top set of menu items dynamic.

**Breaking Changes**

Main menu will be empty after upgrade. To remedy, create Directory Collections to match your needs.

Drops the seldom-used "Do Not Contact (Nationals)" flag for a Member record. If you still need this data, you can create a new Tag for it prior to upgrading.