Simplify Facebook links, create Photo URL field
Prior to this PR, we had to look up the identifier in order to link to a
Facebook account. This isn't user friendly, and impractical for most
purposes since we don't expect to be able to leverage the OpenGraph API
anyway.

The new `photoUrl` will pick up the now-deprecated field to pre-populate
an image for those users, but ultimately it will up to the admins to
provide a photo. Gravatar remains the fall-back option.